add voice class to manage individual sound voices now with bugs

+++ sdl3-sound/src/music.cpp
@@ -8,7 +8,7 @@ const std::vector<Note>& Music::get_notes() const {
     return notes_;
 }
 
-void Music::add_note(int pitch, double duration, double velocity) {
+void Music::add_note(int pitch, float duration, float velocity) {
     notes_.push_back({pitch, duration, velocity});
 }
